fix: note additional licenses for logo and model
Updated the licenses page to note CC licenses for the Chickensoft logo
and mascot model. Also updated the footer link for the licenses page
since CC licenses aren't "open source."

Additionally, markdown processing is currently removing extra line
breaks in plain-text fenced blocks and messing up the paragraphing for
the MIT license. As a workaround, added spaces to empty lines in the
MIT-license fenced block and told VSCode not to remove them.